+++ projects/query-search/src/lib/directives/options-scroll.directive.ts
@@ -21,26 +21,26 @@ export class OptionsScrollDirective implements OnDestroy {
   constructor(public autoComplete: MatAutocomplete) {
     this.autoComplete.opened.pipe(
       tap(() => {
-        // Note: When autocomplete raises opened, panel is not yet created (by Overlay)
-        // Note: The panel will be available on next tick
-        // Note: The panel wil NOT open if there are no options to display
         setTimeout(() => {
-          // Note: remove listner just for safety, in case the close event is skipped.
           this.removeScrollEventListener();
-          this.autoComplete.panel.nativeElement
-            .addEventListener('scroll', this.onScroll.bind(this))
+          if (!!this.autoComplete.panel && !!this.autoComplete.panel.nativeElement) {
+            this.autoComplete.panel.nativeElement.addEventListener('scroll', this.onScroll.bind(this))
+          }
         });
       }),
-      takeUntil(this._onDestroy)).subscribe();
+      takeUntil(this._onDestroy)
+    ).subscribe();
 
     this.autoComplete.closed.pipe(
       tap(() => this.removeScrollEventListener()),
-      takeUntil(this._onDestroy)).subscribe();
+      takeUntil(this._onDestroy)
+    ).subscribe();
   }
 
   private removeScrollEventListener() {
-    this.autoComplete.panel.nativeElement
-      .removeEventListener('scroll', this.onScroll);
+    if (!!this.autoComplete.panel && !!this.autoComplete.panel.nativeElement) {
+      this.autoComplete.panel.nativeElement.removeEventListener('scroll', this.onScroll);
+    }
   }
 
   ngOnDestroy() {
@@ -58,9 +58,7 @@ export class OptionsScrollDirective implements OnDestroy {
       const threshold = this.thresholdPercent * 100 * target.scrollHeight / 100;
       const current = target.scrollTop + target.clientHeight;
 
-      //console.log(`scroll ${current}, threshold: ${threshold}`)
       if (current > threshold) {
-        //console.log('load next page');
         this.scroll.next({autoComplete: this.autoComplete, scrollEvent: event});
       }
     }